+++ Opcodes/sockrecv.c
@@ -518,7 +518,7 @@ static int32_t init_srecv(CSOUND *csound, SOCKRECVT *p)
 #if defined(WIN32) && !defined(__CYGWIN__)
     if (p->sock == SOCKET_ERROR) {
       err = WSAGetLastError();
-      csound->InitError(csound, Str("socket failed with error: %ld\n"), err);
+      csound->InitError(csound, Str("socket failed with error: %ld\n"), (long)err);
     }
 #else
     if (UNLIKELY(p->sock < 0)) {
